Move removeDisallowedHtmlContent to brambleUtils; define regex in WebLab
